fix: update data types for MinValue and MaxValue in DTOs
- Changed `MinValue` and `MaxValue` properties to `decimal?` type in `CreateContentTypeFieldDto` to match the `Field` entity.
- Updated `FieldDto` to change `MinValue` and `MaxValue` properties to `decimal?` for consistency.
- Added `IsUnique`, `MinDateTime`, and `MaxDateTime` properties to `CreateContentTypeFieldDto` for additional field constraints.

+++ AlturaCMS.Application/Features/Fields/Queries/GetFields/GetFieldsResponse.cs
@@ -14,8 +14,8 @@ public class FieldDto
     public bool IsRequired { get; set; }
     public int? MinLength { get; set; }
     public int? MaxLength { get; set; }
-    public int? MinValue { get; set; }
-    public int? MaxValue { get; set; }
+    public decimal? MinValue { get; set; }
+    public decimal? MaxValue { get; set; }
     public string? RegexPattern { get; set; } = string.Empty;
     public List<string>? AllowedValues { get; set; } = [];
 }
